SE-mobile-tests/requirements.txt
pytest>=8.0.0
playwright>=1.42.0
allure-pytest>=2.13.5
pytest-rerunfailures>=14.0
pytest-timeout>=2.3.1
